Remove jcat_context_sign()
It doesn't logically make sense; what are we signing with if not using the
client cert? If we're using the client cert, then only one engine can possibly
be supported if we're returning just one thing.

This also forces us to return the JcatBlob from the signing operation, which
is logically also correct; only the specific engine knows the correct flags to
use when exporting, e.g. `JCAT_BLOB_FLAG_IS_UTF8`.