build: Require Python explicitly
When the first argument to python.find_installation is omitted,
Meson will return the Python instance used to run it.
But that can be a completely different installation than the one on PATH.

This change was initially introduced in
https://github.com/hughsie/libjcat/commit/19af407b582cb93329ef2d8f7fb447230c0f5f58
but reverted in https://github.com/hughsie/libjcat/commit/aa20cd6c092eac6d6bc79f09cd73ad13999cf840
due to concern about FreeBSD 12.2.
But FreeBSD 12.2 installs python3 on PATH nowadays and even if it did not,
the inability to find Python is a bug to fix in Meson.